requirements.txt
Django==1.11.3
djangorestframework==3.6.3
drfdocs==0.0.11
pytz==2017.2
